Clean up some deprecated type in our npm package (#23044)
These are things like `DatabaseReader` which have been replaced with `GenericDatabaseReader` (and the codegen-ed data model specific `DatabaseReader`).

This is a breaking change, but one I believe we should be fine with (some TS users will have to update their types as part of upgrading, the types have been marked deprecated for several months).

GitOrigin-RevId: 4326460172b5f8435a3e567f01ac57a2d1206c1b