Try adding some examples of helper functions to our best practices (#24644)
We sometimes get the question of "how do I call a mutation from another mutation" or see folks using `ctx.runAction` as opposed to using a helper function.

I wanted to try adding an example of helper functions and to also touch on the advantages of doing this over alternatives like calling two separate mutations (either from the client or via scheduling) or using `ctx.runAction`.

GitOrigin-RevId: 4bfa9e0bb4ac236b1c43c106ed224fef60c171da