Add a "Production Checklist" page to the Convex docs (#32813)
This page contains a number of anti-patterns we've noticed ourselves in bigger codebases. Some of these can pretty easily turn into lint rules or warnings we emit in the Function Health page on the dashboard, but I want to start by documenting these.

My aim is for all of these to be fairly easy to identify and quick to fix (some light to moderate refactoring) and for them to all have a clear benefit.

Edit: Between the initial version of this PR and now, I made this page replace the existing "Best practices" page vs. adding it as an additional page.
GitOrigin-RevId: be8793192d53ec6016c70f9ec54cf7544f91abab